requirements_backup.txt
streamlit
google-generativeai
python-dotenv
pdf2imagepip
pdf2image